Merge '2.12.0-259.0.dev' into beta

+++ pkg/front_end/test/crashing_test_case_minimizer_impl.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:async' show Future, StreamSubscription;
 
 import 'dart:convert' show JsonEncoder, jsonDecode, utf8;
@@ -51,7 +53,7 @@ import 'package:front_end/src/fasta/util/direct_parser_ast_helper.dart';
 import 'package:front_end/src/fasta/util/textual_outline.dart'
     show textualOutline;
 
-import 'package:kernel/ast.dart' show Component;
+import 'package:kernel/ast.dart' show Component, LibraryPart;
 
 import 'package:kernel/binary/ast_to_binary.dart' show BinaryPrinter;
 
@@ -85,12 +87,14 @@ class TestMinimizerSettings {
   bool widgetTransformation = false;
   final List<Uri> invalidate = [];
   String targetString = "VM";
+  bool noTryToDeleteEmptyFilesUpFront = false;
   bool oldBlockDelete = false;
   bool lineDelete = false;
   bool byteDelete = false;
   bool askAboutRedirectCrashTarget = false;
   bool autoUncoverAllCrashes = false;
   int stackTraceMatches = 1;
+  String lookForErrorErrorOnReload;
   final Set<String> askedAboutRedirect = {};
   final List<Map<String, dynamic>> fileSystems = [];
   final Set<String> allAutoRedirects = {};
@@ -114,6 +118,7 @@ class TestMinimizerSettings {
       'widgetTransformation': widgetTransformation,
       'invalidate': invalidate.map((uri) => uri.toString()).toList(),
       'targetString': targetString,
+      'noTryToDeleteEmptyFilesUpFront': noTryToDeleteEmptyFilesUpFront,
       'oldBlockDelete': oldBlockDelete,
       'lineDelete': lineDelete,
       'byteDelete': byteDelete,
@@ -123,6 +128,7 @@ class TestMinimizerSettings {
       'askedAboutRedirect': askedAboutRedirect.toList(),
       'fileSystems': fileSystems,
       'allAutoRedirects': allAutoRedirects.toList(),
+      'lookForErrorErrorOnReload': lookForErrorErrorOnReload,
     };
   }
 
@@ -137,6 +143,7 @@ class TestMinimizerSettings {
     invalidate.addAll(
         (json["invalidate"] as List).map((uriString) => Uri.parse(uriString)));
     targetString = json["targetString"];
+    noTryToDeleteEmptyFilesUpFront = json["noTryToDeleteEmptyFilesUpFront"];
     oldBlockDelete = json["oldBlockDelete"];
     lineDelete = json["lineDelete"];
     byteDelete = json["byteDelete"];
@@ -149,10 +156,18 @@ class TestMinimizerSettings {
     fileSystems.addAll((json["fileSystems"] as List).cast());
     allAutoRedirects.clear();
     allAutoRedirects.addAll((json["allAutoRedirects"] as List).cast());
+    lookForErrorErrorOnReload = json["lookForErrorErrorOnReload"];
 
     _fsNotInitial.initializeFromJson(fileSystems.removeLast());
   }
 }
+
+// TODO(jensj): The different cuts and inlines in this file aren't tested.
+// The probably should be. So they should probably be factored out so they can
+// be tested and then tested.
+// Similarly the whole +/- 1 thing to cut out what we want is weird and should
+// be factored out into helpers too, or we should have some sort of visitor
+// that can include the "real" numbers (which should then also be tested).
 
 class TestMinimizer {
   final TestMinimizerSettings _settings;
@@ -163,6 +178,7 @@ class TestMinimizer {
   bool _skip = false;
   bool _check = false;
   int _currentFsNum = -1;
+  bool _gotWantedError = false;
 
   Component _latestComponent;
   IncrementalCompiler _latestCrashingIncrementalCompiler;
@@ -177,7 +193,8 @@ class TestMinimizer {
       stdin.echoMode = false;
       stdin.lineMode = false;
     } catch (e) {
-      print("error setting settings on stdin");
+      print("Trying to setup 'stdin' failed. Continuing anyway, "
+          "but 'q', 'i' etc might not work.");
     }
     _stdinSubscription = stdin.listen((List<int> event) {
       if (event.length == 1 && event.single == "q".codeUnits.single) {
@@ -322,20 +339,22 @@ class TestMinimizer {
       }
 
       // Try to delete empty files.
-      bool changedSome2 = true;
-      while (changedSome2) {
-        if (await _shouldQuit()) break;
-        changedSome2 = false;
-        for (int i = 0; i < uris.length; i++) {
+      if (!_settings.noTryToDeleteEmptyFilesUpFront) {
+        bool changedSome2 = true;
+        while (changedSome2) {
           if (await _shouldQuit()) break;
-          Uri uri = uris[i];
-          if (_fs.data[uri] == null || _fs.data[uri].isNotEmpty) continue;
-          print("About to work on file $i of ${uris.length}");
-          await _deleteContent(uris, i, false, initialComponent,
-              deleteFile: true);
-          if (_fs.data[uri] == null) {
-            changedSome = true;
-            changedSome2 = true;
+          changedSome2 = false;
+          for (int i = 0; i < uris.length; i++) {
+            if (await _shouldQuit()) break;
+            Uri uri = uris[i];
+            if (_fs.data[uri] == null || _fs.data[uri].isNotEmpty) continue;
+            print("About to work on file $i of ${uris.length}");
+            await _deleteContent(uris, i, false, initialComponent,
+                deleteFile: true);
+            if (_fs.data[uri] == null) {
+              changedSome = true;
+              changedSome2 = true;
+            }
           }
         }
       }
@@ -425,6 +444,15 @@ class TestMinimizer {
           }
         }
       }
+
+      // Partially a sanity-check, and partially to make sure there's a latest
+      // component for [_tryToRemoveUnreferencedFileContent] to work with (if
+      // crashing after an invalidation).
+      if (!await _crashesOnCompile(initialComponent)) {
+        throw "At a state where we should crash, we didn't!";
+      }
+      await _tryToRemoveUnreferencedFileContent(initialComponent,
+          deleteFile: true);
     }
 
     if (await _shouldQuit()) {
@@ -544,6 +572,10 @@ class TestMinimizer {
           _replaceRange(replacements, originalBytes);
 
       // Step 2: Find the last import/export.
+      // TODO(jensj): This doesn't work if
+      // * The file we're inlining into doesn't have any imports/exports but do
+      //   have a `library` declaration.
+      // * The file we're inlining has a library declaration.
       int offsetOfLast = 0;
       ast = getAST(withoutInlineable,
           includeBody: false,
@@ -689,20 +721,20 @@ class TestMinimizer {
 
     // TODO(jensj): don't use full uris.
     print("""
-        # Copyright (c) 2020, the Dart project authors. Please see the AUTHORS file
-        # for details. All rights reserved. Use of this source code is governed by a
-        # BSD-style license that can be found in the LICENSE.md file.
-        
-        # Reproduce a crash.
-        
-        type: newworld""");
+# Copyright (c) 2021, the Dart project authors. Please see the AUTHORS file
+# for details. All rights reserved. Use of this source code is governed by a
+# BSD-style license that can be found in the LICENSE.md file.
+
+# Reproduce a crash.
+
+type: newworld""");
     if (_settings.widgetTransformation) {
       print("trackWidgetCreation: true");
       print("target: DDC # basically needed for widget creation to be run");
     }
     print("""
-        worlds:
-          - entry: $_mainUri""");
+worlds:
+  - entry: $_mainUri""");
     if (_settings.experimentalInvalidation) {
       print("    experiments: alternative-invalidation-strategy");
     }
@@ -876,9 +908,11 @@ class TestMinimizer {
     bool removedSome = false;
     if (await _shouldQuit()) return;
     for (MapEntry<Uri, Uint8List> entry in _fs.data.entries) {
-      if (entry.value == null || entry.value.isEmpty) continue;
+      if (entry.value == null) continue;
+      if (!deleteFile && entry.value.isEmpty) continue;
       if (!entry.key.toString().endsWith(".dart")) continue;
-      if (!neededUris.contains(entry.key) && _fs.data[entry.key].length != 0) {
+      if (!neededUris.contains(entry.key) &&
+          (deleteFile || _fs.data[entry.key].length != 0)) {
         if (deleteFile) {
           _fs.data[entry.key] = null;
         } else {
@@ -1423,6 +1457,10 @@ class TestMinimizer {
             }
 
             // Try to remove "extends", "implements" etc.
+            // TODO(jensj): The below removes from the "extends" (etc) until
+            // (but excluding) the "{". This should be improved so it can remove
+            // _one_ part at a time. E:g. if it says "class A implements B, C {"
+            // we could try to remove "B, " or ", C" etc.
             if (decl.getClassExtends().extendsKeyword != null) {
               helper.replacements.add(new _Replacement(
                   decl.getClassExtends().extendsKeyword.offset - 1,
@@ -1728,8 +1766,9 @@ class TestMinimizer {
   }
 
   bool _isUriNnbd(Uri uri) {
+    Uri asImportUri = _getImportUri(uri);
     LibraryBuilder libraryBuilder = _latestCrashingIncrementalCompiler
-        .userCode.loader.builders[_getImportUri(uri)];
+        .userCode.loader.builders[asImportUri];
     if (libraryBuilder != null) {
       return libraryBuilder.isNonNullableByDefault;
     }
@@ -1737,17 +1776,25 @@ class TestMinimizer {
     for (LibraryBuilder libraryBuilder
         in _latestCrashingIncrementalCompiler.userCode.loader.builders.values) {
       if (libraryBuilder.importUri == uri) {
-        print("Found $uri as ${libraryBuilder.importUri} "
-            "(!= ${_getImportUri(uri)})");
+        print("Found $uri as ${libraryBuilder.importUri} (!= ${asImportUri})");
         return libraryBuilder.isNonNullableByDefault;
       }
+      // Check parts too.
+      for (LibraryPart part in libraryBuilder.library.parts) {
+        Uri thisPartUri = libraryBuilder.importUri.resolve(part.partUri);
+        if (thisPartUri == uri || thisPartUri == asImportUri) {
+          print("Found $uri as part of ${libraryBuilder.importUri}");
+          return libraryBuilder.isNonNullableByDefault;
+        }
+      }
     }
-    // This might be parts?
     throw "Couldn't lookup $uri at all!";
   }
 
   Future<bool> _crashesOnCompile(Component initialComponent) async {
     IncrementalCompiler incrementalCompiler;
+    _gotWantedError = false;
+    bool didNotGetWantedErrorAfterFirstCompile = true;
     if (_settings.noPlatform) {
       incrementalCompiler = new IncrementalCompiler(_setupCompilerContext());
     } else {
@@ -1758,20 +1805,33 @@ class TestMinimizer {
     try {
       _latestComponent = await incrementalCompiler.computeDelta();
       if (_settings.serialize) {
+        // We're asked to serialize, probably because it crashes in
+        // serialization.
         ByteSink sink = new ByteSink();
         BinaryPrinter printer = new BinaryPrinter(sink);
         printer.writeComponentFile(_latestComponent);
-        sink.builder.takeBytes();
       }
+
+      if (_gotWantedError) didNotGetWantedErrorAfterFirstCompile = false;
+
       for (Uri uri in _settings.invalidate) {
         incrementalCompiler.invalidate(uri);
         Component delta = await incrementalCompiler.computeDelta();
         if (_settings.serialize) {
+          // We're asked to serialize, probably because it crashes in
+          // serialization.
           ByteSink sink = new ByteSink();
           BinaryPrinter printer = new BinaryPrinter(sink);
           printer.writeComponentFile(delta);
           sink.builder.takeBytes();
         }
+      }
+      if (_settings.lookForErrorErrorOnReload != null &&
+          _gotWantedError &&
+          didNotGetWantedErrorAfterFirstCompile) {
+        // We throw here so the "compile crashes"... This is looking for
+        // crashes after-all :)
+        throw "Got the wanted error!";
       }
       _latestComponent = null; // if it didn't crash this isn't relevant.
       return false;
@@ -1879,6 +1939,12 @@ class TestMinimizer {
     options.omitPlatform = false;
     options.onDiagnostic = (DiagnosticMessage message) {
       // don't care.
+      // Except if we're looking to trigger a specific error on reload.
+      if (_settings.lookForErrorErrorOnReload != null &&
+          message.ansiFormatted.first
+              .contains(_settings.lookForErrorErrorOnReload)) {
+        _gotWantedError = true;
+      }
     };
     if (_settings.noPlatform) {
       options.librariesSpecificationUri = null;
